VertexArray

+++ Capybara/src/Capybara/Renderer/Buffer.h
@@ -32,7 +32,7 @@ namespace Capybara
         std::string Name;
         ShaderDataType Type;
         uint32_t Size;
-        uint32_t Offset;
+        uint64_t Offset;
         bool Normalized;
         BufferElement() {}
         BufferElement(ShaderDataType type, const std::string& name, bool normalized = false)
@@ -71,7 +71,8 @@ namespace Capybara
             CalculateOffsetAndStride();
         }
         inline uint32_t GetStride() const { return m_Stride; }
-        inline const std::vector<BufferElement>& GetElements() const {return m_Elements;}
+        inline const std::vector<BufferElement>& GetElements() const { return m_Elements; }
+        
         std::vector<BufferElement>::iterator begin() { return m_Elements.begin(); }
         std::vector<BufferElement>::iterator end() { return m_Elements.end(); }
         std::vector<BufferElement>::const_iterator begin() const { return m_Elements.begin(); }
@@ -108,6 +109,7 @@ namespace Capybara
 
         static VertexBuffer* Create(float* vertices, uint32_t size);
     };
+    
     class IndexBuffer
     {
     public:
